requirements.txt
Keras==2.4.3
opencv-python-headless==4.5.1.48
cvlib==0.2.4
numpy==1.19.5
streamlit==0.79.0
Pillow==8.0.0
tensorflow==2.0.0
requests==2.25.1